+++ src/stac_auth_proxy/app.py
@@ -56,6 +56,22 @@ def create_app(settings: Optional[Settings] = None) -> FastAPI:
         oidc_config_url=str(settings.oidc_discovery_url),
     )
 
+    # TODO: How can we inject the collections_filter into only the endpoints that need it?
+    # for endpoint, methods in settings.collections_filter_endpoints.items():
+    #     app.add_api_route(
+    #         endpoint,
+    #         partial(
+    #             proxy_handler.stream, collections_filter=settings.collections_filter
+    #         ),
+    #         methods=methods,
+    #         dependencies=[Security(auth_scheme.maybe_validated_user)],
+    #     )
+
+    # skip = [
+    #     settings.openapi_spec_endpoint,
+    #     *settings.collections_filter_endpoints,
+    # ]
+
     # Endpoints that are explicitely marked private
     for path, methods in settings.private_endpoints.items():
         app.add_api_route(
